jit: add jit_if_available option
This commit adds a new builder option, `jit_if_available`, that is like
the existing `jit` option, but will silently fall back to the non-JIT
matcher if JIT compilation fails (and will also emit a debug log
message). This is useful for platforms that aren't supported by PCRE2's
JIT.

We also add a new `is_jit_enabled` public API function, which permits
users to query whether PCRE2's JIT is enabled or not. We use this
routine in a test to check the semantics of the `jit` builder option.

Fixes #3