jit: expose an option for increasing JIT stack size
This pretty much does what it says on the tin. The default maximum size
is only 32KB, and there have been some reports in the wild that run afoul
of this limit. So we permit increasing it.

This required using a match context. So we setup a MatchConfig for future
options (such as controlling depth/heap/match limit).