sys: disable jit on android, aarch64 and armv7
Apparently it isn't building on these targets?

My best guess is that there is something wrong I'm doing in the build
script when compiling for PCRE2. And probably that should be fixed. I'm
pretty sure, for example, PCRE2's JIT supports arm. But this seems like
a better choice to make the build succeed.

Closes #25, Closes #29, Closes #30