Expose the PCRE2_NEVER_UTF flag
By default, PCRE2 enables the strange sequence "(*UTF)" which turns on UTF
validity checking for both patterns and subjects. This is hinted at as a
potential security concern in the man page:

"If the data string is very long, such a check might use sufficiently many
resources as to cause your application to lose performance"

For this reason, pcre2 provides a flag to avoid interpreting this
sequence. Re-expose that in rust-pcre2, under the clearer name
`block_utf_pattern_directive`.