git: remove pcre2-sys/pcre2 submodule
I fucking hate submodules. So I'm going to switch to a lower-tech
solution that will just copy the files into the repository.

Every time I come into this repo to update PCRE2, I have to do a deep
dive on submodules to figure out how to do a simple update. No more. I'm
done.

This commit breaks the PCRE2_SYS_STATIC=1 build option, but we'll fix it
up in a subsequent commit.