sys: update PCRE2 static build infrastructure
We still retain the common/default case of "dynamically link to PCRE2
system library" if possible, but otherwise update the build to be a bit
simpler for the static build. We also add a shell script for pulling in
a PCRE2 release into this repository (which we will add in a subsequent
commit).

This infrastructure was copied from the (yet to be released) rebar
project.

Fixes #24 (hopefully)