find_iter, captures_iter: fix empty matches in UTF mode
Following an empty match, the iterator (Matches or CaptureMatches)
advances the last_end position so as not to return the same match
twice.

However, if the regex uses UTF mode, the position passed to
find_at_with_match_data or captures_read_at is required to be a UTF-8
character boundary, so last_end must be advanced by a whole UTF-8
character, not just one byte.

Determining whether or not the regex is using UTF mode requires
PCRE2_INFO_ALLOPTIONS (checking the config is not enough.)