bytes: avoid `CaptureLocations::get()` integer overflow panics
This also impacts `Captures::get()`.

PCRE2 keeps the ovector index as a 16-bit unsigned integer, but the
external API allows values up to `usize::MAX`, which would cause
panics when converted into a match pair. This means that a panic
would only occur prior to this commit *only* when the index given was
incorrect/impossible. Still, we shouldn't panic in such cases.

Fixes #50